requirements.txt
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
fonttools==4.49.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
